requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
pinecone>=8.0.0
sentence-transformers>=3.0.0
pymupdf>=1.25.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
httpx>=0.28.0
pydantic>=2.10.0
